delete old versions of controllers and rename new versions to replace them
